Add files via upload
This is an MS Excel Report for a clothing store from which insights can be gathered.
I cleaned the original dataset, made it consistent, and added a few columns to make analysis easier. Then, I create reports using pivot table table.
All the reports can be seen in a dashboard-type format in the Store Reports Dashboard sheet in this excel file.